Simplify AbstractColumnReader by reducing fields and methods in it
